not working: red lines appear when crossing other lines
